catalog-tovar adapt - finished
